refactor: Merge types/docs from testing/auth/oauth_scope_validator.pyi
- Merged type hints and docstrings from crudclient/testing/auth/oauth_scope_validator.pyi into crudclient/testing/auth/oauth_scope_validator.py.

- Deleted crudclient/testing/auth/oauth_scope_validator.pyi.
